+++ x/auth/types/expected_keepers.go
@@ -3,7 +3,7 @@ package types
 import (
 	"context"
 
-	"google.golang.org/protobuf/runtime/protoiface"
+	"cosmossdk.io/core/transaction"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
@@ -17,7 +17,7 @@ type BankKeeper interface {
 
 // AccountsModKeeper defines the contract for x/accounts APIs
 type AccountsModKeeper interface {
-	SendModuleMessageUntyped(ctx context.Context, sender []byte, msg protoiface.MessageV1) (protoiface.MessageV1, error)
+	SendModuleMessage(ctx context.Context, sender []byte, msg transaction.Msg) (transaction.Msg, error)
 	IsAccountsModuleAccount(ctx context.Context, accountAddr []byte) bool
 	NextAccountNumber(ctx context.Context) (accNum uint64, err error)